# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "bro back swagger",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8000/api",
      "description": "local server"
    }
  ],
  "security": [
    {
      "JWT": []
    }
  ],
  "tags": [
    {
      "name": "Auth operations"
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "description": "register endpoint to create profile",
        "operationId": "registerAuth",
        "tags": ["Auth operations"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthLogin"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "register successfulled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Auth"                  
                }
              }
            }
          },
          "400": {
            "description": "parameter is missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/login": {
      "post": {
        "description": "login endpoint to get token",
        "operationId": "loginAuth",
        "tags": ["Auth operations"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthInfo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "login successfulled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Auth"                  
                }
              }
            }
          },
          "400": {
            "description": "parameter is missing",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/auth/current": {
      "get": {
        "description": "get current logged user infos",
        "operationId": "currentAuth",
        "tags": ["Auth operations"],
        "responses": {
          "200": {
            "description": "user info were obtained",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Auth"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorize request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "userLogin": {
        "type": "object",
        "properties": {
          "pseudo": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "authUser": {
        "type": "object",
        "properties": {
          "pseudo": {
            "type": "string"
          },
          "token": {
            "type": "string"
          }
        }
      },
      "Auth": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/authUser"
          }
        }
      },
      "AuthInfo": {
        "type": "object",
        "properties": {
          "user": {
            "$ref": "#/components/schemas/userLogin"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "errors": {
            "type": "object"
          }
        }
      }
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": ""
      }
    }
  }
}